Moved each specific updateDiscount code into its own method in the DiscountOffer prototype.
I a not very pleased with the code, however I didnt find a nice way to write these custom behaviors without breaking the current API.

One option would have been to create a specific class for each Offer that deviates from the default behavior. Each of this class would have extended the DiscountOffer class so that we would only have to override the methods we needed (updateDiscount or updateExpiration).

One other option (also API breaking) would have been for the DiscountOffer constructor to accept as arguments (optionnaly) an updateDiscount function and an updateExpiration function.
That way we could specify how to update the discount when creating our different offers.